demo: fix 'mini_c' compile errors with GCC 8.3.0
Fixes #3341

+++ repos/demo/src/lib/mini_c/malloc_free.cc
@@ -32,7 +32,7 @@ static Allocator &alloc()
 void mini_c_init(Allocator &alloc) { _alloc_ptr = &alloc; }
 
 
-extern "C" void *malloc(unsigned size)
+extern "C" void *malloc(size_t size)
 {
 	/*
 	 * We store the size of the allocation at the very
@@ -50,7 +50,7 @@ extern "C" void *malloc(unsigned size)
 }
 
 
-extern "C" void *calloc(unsigned nmemb, unsigned size)
+extern "C" void *calloc(size_t nmemb, size_t size)
 {
 	void *addr = malloc(nmemb*size);
 	memset(addr, 0, nmemb*size);
